+++ lib/src/wallet/wallet_store.dart
@@ -5,7 +5,6 @@ import 'package:base_codecs/base_codecs.dart';
 import 'package:bip32/bip32.dart';
 import 'package:bip39/bip39.dart';
 import 'package:crypto/crypto.dart';
-import 'package:dart_ssi/didcomm.dart';
 import 'package:dart_web3/credentials.dart';
 import 'package:dart_web3/crypto.dart';
 import 'package:ed25519_edwards/ed25519_edwards.dart' as ed;
@@ -13,6 +12,7 @@ import 'package:ed25519_hd_key/ed25519_hd_key.dart';
 import 'package:hive/hive.dart';
 import 'package:x25519/x25519.dart' as x;
 
+import '../../didcomm.dart';
 import '../util/private_util.dart';
 import '../util/utils.dart';
 import 'hive_model.dart';
@@ -792,8 +792,8 @@ class WalletStore {
   }
 
   /// Stores a configuration Entry.
-  void storeConfigEntry(String key, String value) {
-    _configBox!.put(key, value);
+  Future<void> storeConfigEntry(String key, String value) async {
+    await _configBox!.put(key, value);
   }
 
   /// Returns the configuration Entry for [key].